+++ src/util/packed/monotonic_block_packed_reader.rs
@@ -188,11 +188,11 @@ impl Display for MonotonicBlockPackedReader {
     }
 }
 impl Accountable for MonotonicBlockPackedReader {
-    fn ram_bytes_used(&self) -> u64 {
+    fn ram_bytes_used(&self) -> i64 {
         let mut size_in_bytes = 0;
         size_in_bytes += RamUsageEstimator::size_of_vec(&self.min_values);
         size_in_bytes += RamUsageEstimator::size_of_vec(&self.averages);
-        size_in_bytes += self.total_byte_count as u64;
+        size_in_bytes += self.total_byte_count;
         size_in_bytes
     }
 }
